Move reducers for failures
