+++ frontend/src/reducers/failures/details/index.js
