+++ frontend/src/reducers/failures/index.spec.js
